# docs/openapi/private-backend.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Me 2 Backend API",
    "description": "Backend Server API endpoints for Me 2 - Node.js/Express",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/profile-views/{userId}": {
      "get": {
        "tags": ["Public", "Private"],
        "summary": "GET /api/profile-views/{userId}",
        "description": "/api/profile-views/{userId}\nGet the profile view for a user (composite view with sections and optionally memories)",
        "operationId": "GetProfileView",
        "parameters": [
          {
            "description": "The ID of the user",
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile view data with sections and optionally memories",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "404": {
            "description": "Profile not found",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": {} }
            }
          }
        },
        "security": []
      }
    },
    "/api/profile/generate": {
      "post": {
        "tags": ["Private"],
        "summary": "POST /api/profile/generate",
        "description": "/api/profile/generate\nGenerate a user profile using AI and save to database",
        "operationId": "GenerateProfile",
        "parameters": [],
        "requestBody": {
          "description": "Generation request with optional userId",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateProfileRequest",
                "description": "Generation request with optional userId"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated profile metadata",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "500": {
            "description": "AI service or database error",
            "content": {
              "application/json": { "schema": {} }
            }
          }
        },
        "security": []
      }
    },
    "/api/profile/trigger-auto-update": {
      "post": {
        "tags": ["Private"],
        "summary": "POST /api/profile/trigger-auto-update",
        "description": "/api/profile/trigger-auto-update\nTrigger profile auto-update after memory operations",
        "operationId": "TriggerAutoUpdate",
        "parameters": [],
        "requestBody": {
          "description": "Auto-update request with userId and successCount",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerAutoUpdateRequest",
                "description": "Auto-update request with userId and successCount"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Auto-update result with triggered status",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "400": {
            "description": "Missing required fields",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": {} }
            }
          }
        },
        "security": []
      }
    },
    "/api/profile/g-eval": {
      "post": {
        "tags": ["Private"],
        "summary": "POST /api/profile/g-eval",
        "description": "/api/profile/g-eval\nGet g-eval score from AI server",
        "operationId": "GetGEval",
        "parameters": [],
        "requestBody": {
          "description": "G-eval request with userId",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GEvalRequest",
                "description": "G-eval request with userId"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "G-eval score",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": {} }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateProfileRequest": {
        "properties": {
          "userId": { "type": "string" }
        },
        "type": "object",
        "additionalProperties": true
      },
      "TriggerAutoUpdateRequest": {
        "properties": {
          "userId": { "type": "string" },
          "successCount": { "type": "number", "format": "double" }
        },
        "required": ["userId", "successCount"],
        "type": "object",
        "additionalProperties": true
      },
      "GEvalRequest": {
        "properties": {
          "userId": { "type": "string" }
        },
        "required": ["userId"],
        "type": "object",
        "additionalProperties": true
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [{ "url": "/" }]
}